+++ CHANGELOG.md
@@ -1,3 +1,9 @@
-## [0.0.1] - TODO: Add release date.
 
-* TODO: Describe initial release.
+## [0.0.2]
+
+* Added entries to return iterable of cache elements.
+* Renamed former entries to length for consistency. 
+
+## [0.0.1]
+
+* Initial cache implementation.